implemented algorithm for matching strip centroids with wire hits

+++ libraries/FDC/DFDCPseudo.h
@@ -14,8 +14,8 @@
 #include <sstream>
 
 typedef struct {
-  double pos;
-  double q;
+  float pos;
+  float q;
 }centroid_t;
